Move model tests to the relevant package that matches the source
