+++ ryft-ui/src/test/java/com/ryftpay/android/ui/model/googlepay/RyftTokenizationSpecificationTest.kt
@@ -1,4 +1,4 @@
-package com.ryftpay.android.ui.service.request.googlepay
+package com.ryftpay.android.ui.model.googlepay
 
 import com.ryftpay.android.ui.TestData.sandboxPublicApiKey
 import com.ryftpay.android.ui.model.googlepay.RyftTokenizationSpecification
