+++ ryft-ui/src/test/java/com/ryftpay/android/ui/service/request/GooglePayRequestFactoryTest.kt
@@ -1,10 +1,9 @@
-package com.ryftpay.android.ui.service.request.googlepay
+package com.ryftpay.android.ui.service.request
 
 import com.ryftpay.android.ui.TestData.merchantInfo
 import com.ryftpay.android.ui.TestData.ryftTokenizationSpecification
 import com.ryftpay.android.ui.TestData.transactionInfo
 import com.ryftpay.android.ui.model.googlepay.LoadPaymentDataRequest
-import com.ryftpay.android.ui.service.request.GooglePayRequestFactory
 import org.amshove.kluent.shouldBeEqualTo
 import org.junit.Test
